testscripts/2026.04.01-pagerank-quality-score/requirements.txt
psycopg2-binary
sqlalchemy
networkx
numpy
requests
python-dotenv
